+++ internal/machine/spectrum/bus_test.go
@@ -10,4 +10,6 @@ func TestBus(t *testing.T) {
 	dsl.RunScenarios(t, busScenarios)
 	dsl.RunScenarios(t, keyboardScenarios)
 	dsl.RunScenarios(t, displayScenarios)
+	dsl.RunScenarios(t, machineScenarios)
+	dsl.RunScenarios(t, tapeScenarios)
 }